Refactor user controller, add getUserById method
Refactored the user controller to add a new method, getUserById, which retrieves a user by their ID. This method is used to handle GET requests to the /users/:id endpoint. It fetches the user from the database using the UserService and returns it as a JSON response.

The new method is implemented in the UserController class, and a corresponding route is added in the user.js file in the routes directory.

This commit addresses the need to fetch a specific user by their ID and enhances the functionality of the API.