Refactor User controller: Add createUser method
Update the User controller by adding a new method createUser, responsible for creating a user with the provided user data.

Refactor UserService: Add createUser method

Implement createUser method in the UserService class to handle the creation of a new user. This method creates a User object with the provided user data and saves it to the database.

Update User route: Add POST method for creating a user

In the User route, a new POST method has been added to handle the creation of a user. This method calls the createUser method in the User controller to create a user with the provided data.